+++ libs/agno/agno/models/azure/openai_chat.py
@@ -9,7 +9,7 @@ from agno.models.openai.like import OpenAILike
 try:
     from openai import AsyncAzureOpenAI as AsyncAzureOpenAIClient
     from openai import AzureOpenAI as AzureOpenAIClient
-except (ModuleNotFoundError, ImportError):
+except ImportError:
     raise ImportError("`openai` not installed. Please install using `pip install openai`")
