+++ libs/agno/agno/models/fireworks/fireworks.py
@@ -8,18 +8,18 @@ from agno.models.openai import OpenAILike
 @dataclass
 class Fireworks(OpenAILike):
     """
-    Fireworks model
+    A class for interacting with models hosted on Fireworks.
 
     Attributes:
         id (str): The model name to use. Defaults to "accounts/fireworks/models/llama-v3p1-405b-instruct".
-        name (str): The model name to use. Defaults to "Fireworks: " + id.
+        name (str): The model name to use. Defaults to "Fireworks".
         provider (str): The provider to use. Defaults to "Fireworks".
-        api_key (Optional[str]): The API key to use. Defaults to getenv("FIREWORKS_API_KEY").
+        api_key (Optional[str]): The API key to use.
         base_url (str): The base URL to use. Defaults to "https://api.fireworks.ai/inference/v1".
     """
 
     id: str = "accounts/fireworks/models/llama-v3p1-405b-instruct"
-    name: str = "Fireworks: " + id
+    name: str = "Fireworks"
     provider: str = "Fireworks"
 
     api_key: Optional[str] = getenv("FIREWORKS_API_KEY", None)
