+++ libs/agno/agno/models/mistral/mistral.py
@@ -3,11 +3,11 @@ from os import getenv
 from typing import Any, Dict, Iterator, List, Optional, Union
 
 from agno.exceptions import ModelProviderError
-from agno.media import Image
 from agno.models.base import Model
 from agno.models.message import Message
 from agno.models.response import ModelResponse
-from agno.utils.log import log_error, log_warning
+from agno.utils.log import log_error
+from agno.utils.models.mistral import format_messages
 
 try:
     from mistralai import CompletionEvent
@@ -16,10 +16,8 @@ try:
     from mistralai.models import (
         AssistantMessage,
         HTTPValidationError,
-        ImageURLChunk,
         SDKError,
         SystemMessage,
-        TextChunk,
         ToolMessage,
         UserMessage,
     )
@@ -29,85 +27,8 @@ try:
 
     MistralMessage = Union[UserMessage, AssistantMessage, SystemMessage, ToolMessage]
 
-except (ModuleNotFoundError, ImportError):
+except ImportError:
     raise ImportError("`mistralai` not installed. Please install using `pip install mistralai`")
-
-
-def _format_image_for_message(image: Image) -> Optional[ImageURLChunk]:
-    # Case 1: Image is a URL
-    if image.url is not None:
-        return ImageURLChunk(image_url=image.url)
-    # Case 2: Image is a local file path
-    elif image.filepath is not None:
-        import base64
-        from pathlib import Path
-
-        path = Path(image.filepath) if isinstance(image.filepath, str) else image.filepath
-        if not path.exists() or not path.is_file():
-            log_error(f"Image file not found: {image}")
-            raise FileNotFoundError(f"Image file not found: {image}")
-
-        with open(image.filepath, "rb") as image_file:
-            base64_image = base64.b64encode(image_file.read()).decode("utf-8")
-            return ImageURLChunk(image_url=f"data:image/jpeg;base64,{base64_image}")
-
-    # Case 3: Image is a bytes object
-    elif image.content is not None:
-        import base64
-
-        base64_image = base64.b64encode(image.content).decode("utf-8")
-        return ImageURLChunk(image_url=f"data:image/jpeg;base64,{base64_image}")
-    return None
-
-
-def _format_messages(messages: List[Message]) -> List[MistralMessage]:
-    mistral_messages: List[MistralMessage] = []
-
-    for message in messages:
-        mistral_message: MistralMessage
-        if message.role == "user":
-            if message.audio is not None and len(message.audio) > 0:
-                log_warning("Audio input is currently unsupported.")
-
-            if message.files is not None and len(message.files) > 0:
-                log_warning("File input is currently unsupported.")
-
-            if message.videos is not None and len(message.videos) > 0:
-                log_warning("Video input is currently unsupported.")
-
-            if message.images is not None:
-                content: List[Any] = [TextChunk(type="text", text=message.content)]
-                for image in message.images:
-                    image_content = _format_image_for_message(image)
-                    if image_content:
-                        content.append(image_content)
-                mistral_message = UserMessage(role="user", content=content)
-            else:
-                mistral_message = UserMessage(role="user", content=message.content)
-        elif message.role == "assistant":
-            if message.reasoning_content is not None:
-                mistral_message = UserMessage(role="user", content=message.content)
-            elif message.tool_calls is not None:
-                mistral_message = AssistantMessage(
-                    role="assistant", content=message.content, tool_calls=message.tool_calls
-                )
-            else:
-                mistral_message = AssistantMessage(role=message.role, content=message.content)
-        elif message.role == "system":
-            mistral_message = SystemMessage(role="system", content=message.content)
-        elif message.role == "tool":
-            mistral_message = ToolMessage(name="tool", content=message.content, tool_call_id=message.tool_call_id)
-        else:
-            raise ValueError(f"Unknown role: {message.role}")
-
-        mistral_messages.append(mistral_message)
-
-    # Check if the last message is an assistant message
-    if mistral_messages and hasattr(mistral_messages[-1], "role") and mistral_messages[-1].role == "assistant":
-        # Set prefix=True for the last assistant message to allow it as the last message
-        mistral_messages[-1].prefix = True
-
-    return mistral_messages
 
 
 @dataclass
@@ -115,24 +36,7 @@ class MistralChat(Model):
     """
     MistralChat is a model that uses the Mistral API to generate responses to messages.
 
-    Args:
-        id (str): The ID of the model.
-        name (str): The name of the model.
-        provider (str): The provider of the model.
-        temperature (Optional[float]): The temperature of the model.
-        max_tokens (Optional[int]): The maximum number of tokens to generate.
-        top_p (Optional[float]): The top p of the model.
-        random_seed (Optional[int]): The random seed of the model.
-        safe_mode (bool): The safe mode of the model.
-        safe_prompt (bool): The safe prompt of the model.
-        response_format (Optional[Union[Dict[str, Any], ChatCompletionResponse]]): The response format of the model.
-        request_params (Optional[Dict[str, Any]]): The request parameters of the model.
-        api_key (Optional[str]): The API key of the model.
-        endpoint (Optional[str]): The endpoint of the model.
-        max_retries (Optional[int]): The maximum number of retries of the model.
-        timeout (Optional[int]): The timeout of the model.
-        client_params (Optional[Dict[str, Any]]): The client parameters of the model.
-        mistral_client (Optional[Mistral]): The Mistral client of the model.
+    For more information, see the Mistral API documentation: https://docs.mistral.ai/capabilities/completion/
     """
 
     id: str = "mistral-large-latest"
@@ -263,7 +167,7 @@ class MistralChat(Model):
         Returns:
             ChatCompletionResponse: The response from the model.
         """
-        mistral_messages = _format_messages(messages)
+        mistral_messages = format_messages(messages)
         try:
             response: Union[ChatCompletionResponse, ParsedChatCompletionResponse]
             if self.response_format is not None and self.structured_outputs:
@@ -298,7 +202,7 @@ class MistralChat(Model):
         Returns:
             Iterator[Any]: The streamed response.
         """
-        mistral_messages = _format_messages(messages)
+        mistral_messages = format_messages(messages)
         try:
             stream = self.get_client().chat.stream(
                 model=self.id,
@@ -323,7 +227,7 @@ class MistralChat(Model):
         Returns:
             ChatCompletionResponse: The response from the model.
         """
-        mistral_messages = _format_messages(messages)
+        mistral_messages = format_messages(messages)
         try:
             response: Union[ChatCompletionResponse, ParsedChatCompletionResponse]
             if self.response_format is not None and self.structured_outputs:
@@ -357,7 +261,7 @@ class MistralChat(Model):
         Returns:
             Any: The streamed response.
         """
-        mistral_messages = _format_messages(messages)
+        mistral_messages = format_messages(messages)
         try:
             stream = await self.get_client().chat.stream_async(
                 model=self.id,
