+++ libs/agno/agno/models/perplexity/perplexity.py
@@ -30,8 +30,8 @@ class Perplexity(OpenAILike):
     Attributes:
         id (str): The model id. Defaults to "sonar".
         name (str): The model name. Defaults to "Perplexity".
-        provider (str): The provider name. Defaults to "Perplexity: " + id.
-        api_key (Optional[str]): The API key. Defaults to None.
+        provider (str): The provider name. Defaults to "Perplexity".
+        api_key (Optional[str]): The API key.
         base_url (str): The base URL. Defaults to "https://api.perplexity.ai/chat/completions".
         max_tokens (int): The maximum number of tokens. Defaults to 1024.
     """
