+++ libs/agno/agno/models/sambanova/sambanova.py
@@ -8,14 +8,14 @@ from agno.models.openai.like import OpenAILike
 @dataclass
 class Sambanova(OpenAILike):
     """
-    A class for interacting with Sambanova models.
+    A class for interacting with Sambanova API.
 
     Attributes:
         id (str): The id of the Sambanova model to use. Default is "Meta-Llama-3.1-8B-Instruct".
         name (str): The name of this chat model instance. Default is "Sambanova"
         provider (str): The provider of the model. Default is "Sambanova".
         api_key (str): The api key to authorize request to Sambanova.
-        base_url (str): The base url to which the requests are sent.
+        base_url (str): The base url to which the requests are sent. Defaults to "https://api.sambanova.ai/v1".
     """
 
     id: str = "Meta-Llama-3.1-8B-Instruct"
