+++ libs/agno/agno/models/xai/xai.py
@@ -11,11 +11,11 @@ class xAI(OpenAILike):
     Class for interacting with the xAI API.
 
     Attributes:
-        id (str): The ID of the language model.
-        name (str): The name of the API.
-        provider (str): The provider of the API.
+        id (str): The ID of the language model. Defaults to "grok-beta".
+        name (str): The name of the API. Defaults to "xAI".
+        provider (str): The provider of the API. Defaults to "xAI".
         api_key (Optional[str]): The API key for the xAI API.
-        base_url (Optional[str]): The base URL for the xAI API.
+        base_url (Optional[str]): The base URL for the xAI API. Defaults to "https://api.x.ai/v1".
     """
 
     id: str = "grok-beta"
